add arvnetwork* files, move network code from arvmisc
