+++ src/arvmisc.c
@@ -27,11 +27,6 @@
 #include <stdio.h>
 #include <zlib.h>
 
-#ifdef G_OS_WIN32
-#else
-#include<ifaddrs.h>
-#endif
-
 /**
  * SECTION: arvstatistic
  * @short_description: An histogram tool
@@ -61,77 +56,6 @@ struct _ArvStatistic {
 
 	ArvHistogram *histograms;
 };
-
-#ifdef G_OS_WIN32
-GList* arv_enumerate_network_interfaces(void){
-	#warning NOT YET IMPLEMENTED
-	return NULL;
-}
-
-// http://mingw-users.1079350.n2.nabble.com/IPv6-getaddrinfo-amp-inet-ntop-td5891996.html
-const char *inet_ntop(int af, const void *src, char *dst, socklen_t cnt)
-{
-	if (af == AF_INET)
-	{
-		struct sockaddr_in in;
-		memset(&in, 0, sizeof(in));
-		in.sin_family = AF_INET;
-		memcpy(&in.sin_addr, src, sizeof(struct in_addr));
-		getnameinfo((struct sockaddr *)&in, sizeof(struct sockaddr_in), dst, cnt, NULL, 0, NI_NUMERICHOST);
-		return dst;
-	}
-	else if (af == AF_INET6)
-	{
-		struct sockaddr_in6 in;
-		memset(&in, 0, sizeof(in));
-		in.sin6_family = AF_INET6;
-		memcpy(&in.sin6_addr, src, sizeof(struct in_addr6));
-		getnameinfo((struct sockaddr *)&in, sizeof(struct sockaddr_in6), dst, cnt, NULL, 0, NI_NUMERICHOST);
-		return dst;
-	}
-	return NULL;
-}
-
-#else
-GList*
-arv_enumerate_network_interfaces(void){
-	struct ifaddrs *ifap = NULL;
-	struct ifaddrs *ifap_iter;
-	GList* ret=NULL;
-	if(getifaddrs (&ifap) <0) return NULL;
-	for (ifap_iter = ifap; ifap_iter != NULL; ifap_iter = ifap_iter->ifa_next) {
-		if ((ifap_iter->ifa_flags & IFF_UP) != 0 &&
-			(ifap_iter->ifa_flags & IFF_POINTOPOINT) == 0 &&
-			(ifap_iter->ifa_addr != NULL) &&
-			(ifap_iter->ifa_addr->sa_family == AF_INET)) {
-				ArvIfaceAddr* a=(ArvIfaceAddr*)g_malloc(sizeof(ArvIfaceAddr));
-				memset(a,0,sizeof(ArvIfaceAddr));
-				a->addr=g_memdup(ifap_iter->ifa_addr,sizeof(struct sockaddr));
-				if(ifap_iter->ifa_netmask) a->netmask=g_memdup(ifap_iter->ifa_netmask,sizeof(struct sockaddr));
-				if(ifap_iter->ifa_ifu.ifu_broadaddr) a->broadaddr=g_memdup(ifap_iter->ifa_ifu.ifu_broadaddr,sizeof(struct sockaddr));
-				if(ifap_iter->ifa_name) a->name=g_strdup(ifap_iter->ifa_name);
-				ret=g_list_prepend(ret,a);
-		}
-	}
-	freeifaddrs (ifap);
-	ret=g_list_reverse(ret);
-	return ret;
-};
-#endif
-
-
-void _arv_iface_addr_free(ArvIfaceAddr*);
-
-void _arv_iface_addr_free(ArvIfaceAddr* a){
-	if(a->addr) g_free(a->addr);
-	if(a->netmask) g_free(a->netmask);
-	if(a->broadaddr) g_free(a->broadaddr);
-	if(a->name) g_free(a->name);
-}
-
-void arv_enumerate_network_interfaces_free(GList* ifaces){
-	g_list_free_full(ifaces,(GDestroyNotify)_arv_iface_addr_free);
-}
