+++ src/arvmisc.h
@@ -29,10 +29,6 @@
 
 #include <arvtypes.h>
 
-#ifdef G_OS_WIN32
-// for socklen_t
-#include <gio/gnetworking.h>
-#endif
 
 G_BEGIN_DECLS
 
@@ -76,18 +72,6 @@ ArvPixelFormat 	arv_pixel_format_from_gst_0_10_caps 		(const char *name, int bpp
 
 const char *	arv_vendor_alias_lookup		(const char *vendor);
 
-#define _ARV_IFACES
-
-#ifdef _ARV_IFACES
-GList* arv_enumerate_network_interfaces(void);
-void arv_enumerate_network_interfaces_free(GList*);
-
-#ifdef G_OS_WIN32
-// prototype is missing on mingw, let's reimplement
-const char *inet_ntop(int af, const void *src, char *dst, socklen_t cnt);
-#endif
-
-#endif
 
 G_END_DECLS
